+++ src/modules/block/block_generator.cpp
@@ -292,10 +292,12 @@ block_generator::make_task_config(const model::block_generator_config& config,
 
     if (config.ratio) {
         t_config.synchronizer = &m_synchronizer;
-        m_synchronizer.ratio_reads.store(config.ratio.value().reads,
-                                         std::memory_order_relaxed);
-        m_synchronizer.ratio_writes.store(config.ratio.value().writes,
-                                          std::memory_order_relaxed);
+        m_synchronizer.ratio_reads.store(
+            static_cast<int32_t>(config.ratio.value().reads),
+            std::memory_order_relaxed);
+        m_synchronizer.ratio_writes.store(
+            static_cast<int32_t>(config.ratio.value().writes),
+            std::memory_order_relaxed);
     }
 
     return t_config;
